+++ Source/SwiftLintFramework/Rules/Lint/SuperfluousDisableCommandRule.swift
@@ -9,7 +9,7 @@ public struct SuperfluousDisableCommandRule: ConfigurationProviderRule {
         identifier: "superfluous_disable_command",
         name: "Superfluous Disable Command",
         description: "SwiftLint 'disable' commands are superfluous when the disabled rule would not have " +
-                     "triggered a violation in the disabled region.",
+                     "triggered a violation in the disabled region. Use \" - \" if you wish to document a command.",
         kind: .lint
     )